+++ tests/Controller/CRUDControllerTest.php
@@ -420,12 +420,6 @@ class CRUDControllerTest extends TestCase
         $this->protectedTestedMethods['configure']->invoke($this->controller);
 
         $this->assertSame(123456, $uniqueId);
-
-        $reflector = new \ReflectionObject($this->controller);
-        $attribute = $reflector->getProperty('admin');
-        $attribute->setAccessible(true);
-
-        $this->assertSame($this->admin, $attribute->getValue($this->controller));
     }
 
     public function testConfigureChild(): void
@@ -453,12 +447,6 @@ class CRUDControllerTest extends TestCase
         $this->protectedTestedMethods['configure']->invoke($this->controller);
 
         $this->assertSame(123456, $uniqueId);
-
-        $reflector = new \ReflectionObject($this->controller);
-        $attribute = $reflector->getProperty('admin');
-        $attribute->setAccessible(true);
-
-        $this->assertInstanceOf(\get_class($adminParent), $attribute->getValue($this->controller));
     }
 
     public function testConfigureWithException(): void
@@ -4078,17 +4066,6 @@ class CRUDControllerTest extends TestCase
         $this->assertSame(['flash_batch_delete_success'], $this->session->getFlashBag()->get('sonata_flash_success'));
         $this->assertSame('list', $result->getTargetUrl());
         $this->assertSame('bar', $this->request->request->get('foo'));
-    }
-
-    public function testItThrowsWhenCallingAnUndefinedMethod(): void
-    {
-        $this->expectException(
-            \LogicException::class
-        );
-        $this->expectExceptionMessage(
-            'Call to undefined method Sonata\AdminBundle\Controller\CRUDController::doesNotExist'
-        );
-        $this->controller->doesNotExist();
     }
 
     public function getCsrfProvider()